timer/print_sprite: drop redundant MakeSprite cast, type SpriteGroup.subs as struct Sprite** (byte-identical)
timer FUN_00499500 calls MakeSprite(sprite) (no uint cast, stale TODO removed).
print_sprite SpriteGroup.subs unsigned int*->struct Sprite**; sub-sprite render
calls cast-free, GetVRAMAddress keeps explicit (unsigned int) (numeric +4).

+++ src/legoland/timer.c
@@ -47,6 +47,6 @@ int FUN_00499500(struct Sprite *sprite) {
             return 0;
         }
     }
-    MakeSprite((unsigned int)sprite); /* TODO-fold: timer.c keeps its own struct Sprite view; MakeSprite's canonical param is the uint sprite handle */
+    MakeSprite(sprite);
     return 1;
 }
